refactor(config): extract configuration file reading from resolution
Configuration handling mixed a pure half (defaults, flag precedence,
validation, the attempt-to-model retry ladder) with a single filesystem
read. Move the read into its own module so the pure half can sit in a
pure layer with no filesystem dependency.

Refs #35
